+++ apis/cassandra/v1beta1/webhook_test.go
@@ -318,13 +318,10 @@ func Test_ValidateDatacenterFieldChanges(t *testing.T) {
 					},
 					Racks: []Rack{{
 						Name: "rack0",
-						Zone: "zone0",
 					}, {
 						Name: "rack1",
-						Zone: "zone1",
 					}, {
 						Name: "rack2",
-						Zone: "zone2",
 					}},
 				},
 			},
@@ -348,13 +345,10 @@ func Test_ValidateDatacenterFieldChanges(t *testing.T) {
 					},
 					Racks: []Rack{{
 						Name: "rack0",
-						Zone: "zone0",
 					}, {
 						Name: "rack1",
-						Zone: "zone1",
 					}, {
 						Name: "rack2",
-						Zone: "zone2",
 					}},
 				},
 			},
@@ -505,13 +499,10 @@ func Test_ValidateDatacenterFieldChanges(t *testing.T) {
 				Spec: CassandraDatacenterSpec{
 					Racks: []Rack{{
 						Name: "rack0",
-						Zone: "zone0",
 					}, {
 						Name: "rack1",
-						Zone: "zone1",
 					}, {
 						Name: "rack2",
-						Zone: "zone2",
 					}},
 				},
 			},
@@ -522,10 +513,8 @@ func Test_ValidateDatacenterFieldChanges(t *testing.T) {
 				Spec: CassandraDatacenterSpec{
 					Racks: []Rack{{
 						Name: "rack0",
-						Zone: "zone0",
 					}, {
 						Name: "rack2",
-						Zone: "zone2",
 					}},
 				},
 			},
@@ -540,7 +529,6 @@ func Test_ValidateDatacenterFieldChanges(t *testing.T) {
 				Spec: CassandraDatacenterSpec{
 					Racks: []Rack{{
 						Name: "rack0",
-						Zone: "zone0",
 					}},
 					Size: 6,
 				},
@@ -552,7 +540,6 @@ func Test_ValidateDatacenterFieldChanges(t *testing.T) {
 				Spec: CassandraDatacenterSpec{
 					Racks: []Rack{{
 						Name: "rack0",
-						Zone: "zone0",
 					}},
 					Size: 3,
 				},
@@ -568,13 +555,10 @@ func Test_ValidateDatacenterFieldChanges(t *testing.T) {
 				Spec: CassandraDatacenterSpec{
 					Racks: []Rack{{
 						Name: "rack0",
-						Zone: "zone0",
 					}, {
 						Name: "rack1",
-						Zone: "zone1",
 					}, {
 						Name: "rack2",
-						Zone: "zone2",
 					}},
 				},
 			},
@@ -585,55 +569,14 @@ func Test_ValidateDatacenterFieldChanges(t *testing.T) {
 				Spec: CassandraDatacenterSpec{
 					Racks: []Rack{{
 						Name: "rack0-changed",
-						Zone: "zone0",
 					}, {
 						Name: "rack1",
-						Zone: "zone1",
 					}, {
 						Name: "rack2",
-						Zone: "zone2",
 					}},
 				},
 			},
 			errString: "change rack name from 'rack0' to 'rack0-changed'",
-		},
-		{
-			name: "Changed a rack zone",
-			oldDc: &CassandraDatacenter{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "exampleDC",
-				},
-				Spec: CassandraDatacenterSpec{
-					Racks: []Rack{{
-						Name: "rack0",
-						Zone: "zone0",
-					}, {
-						Name: "rack1",
-						Zone: "zone1",
-					}, {
-						Name: "rack2",
-						Zone: "zone2",
-					}},
-				},
-			},
-			newDc: &CassandraDatacenter{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "exampleDC",
-				},
-				Spec: CassandraDatacenterSpec{
-					Racks: []Rack{{
-						Name: "rack0",
-						Zone: "zone0",
-					}, {
-						Name: "rack1",
-						Zone: "zone1",
-					}, {
-						Name: "rack2",
-						Zone: "zone2-changed",
-					}},
-				},
-			},
-			errString: "change rack zone from 'zone2' to 'zone2-changed'",
 		},
 		{
 			name: "Adding a rack is allowed if size increases",
@@ -645,13 +588,10 @@ func Test_ValidateDatacenterFieldChanges(t *testing.T) {
 					Size: 3,
 					Racks: []Rack{{
 						Name: "rack0",
-						Zone: "zone0",
 					}, {
 						Name: "rack1",
-						Zone: "zone1",
 					}, {
 						Name: "rack2",
-						Zone: "zone2",
 					}},
 				},
 			},
@@ -663,16 +603,12 @@ func Test_ValidateDatacenterFieldChanges(t *testing.T) {
 					Size: 4,
 					Racks: []Rack{{
 						Name: "rack0",
-						Zone: "zone0",
 					}, {
 						Name: "rack1",
-						Zone: "zone1",
 					}, {
 						Name: "rack2",
-						Zone: "zone2",
 					}, {
 						Name: "rack3",
-						Zone: "zone2",
 					}},
 				},
 			},
@@ -687,13 +623,10 @@ func Test_ValidateDatacenterFieldChanges(t *testing.T) {
 				Spec: CassandraDatacenterSpec{
 					Racks: []Rack{{
 						Name: "rack0",
-						Zone: "zone0",
 					}, {
 						Name: "rack1",
-						Zone: "zone1",
 					}, {
 						Name: "rack2",
-						Zone: "zone2",
 					}},
 				},
 			},
@@ -704,16 +637,12 @@ func Test_ValidateDatacenterFieldChanges(t *testing.T) {
 				Spec: CassandraDatacenterSpec{
 					Racks: []Rack{{
 						Name: "rack0",
-						Zone: "zone0",
 					}, {
 						Name: "rack1",
-						Zone: "zone1",
 					}, {
 						Name: "rack2",
-						Zone: "zone2",
 					}, {
 						Name: "rack3",
-						Zone: "zone2",
 					}},
 				},
 			},
@@ -729,10 +658,8 @@ func Test_ValidateDatacenterFieldChanges(t *testing.T) {
 					Size: 9,
 					Racks: []Rack{{
 						Name: "rack0",
-						Zone: "zone0",
 					}, {
 						Name: "rack1",
-						Zone: "zone1",
 					}},
 				},
 			},
@@ -744,13 +671,10 @@ func Test_ValidateDatacenterFieldChanges(t *testing.T) {
 					Size: 11,
 					Racks: []Rack{{
 						Name: "rack0",
-						Zone: "zone0",
 					}, {
 						Name: "rack1",
-						Zone: "zone1",
 					}, {
 						Name: "rack2",
-						Zone: "zone2",
 					}},
 				},
 			},
@@ -766,10 +690,8 @@ func Test_ValidateDatacenterFieldChanges(t *testing.T) {
 					Size: 9,
 					Racks: []Rack{{
 						Name: "rack0",
-						Zone: "zone0",
 					}, {
 						Name: "rack1",
-						Zone: "zone1",
 					}},
 				},
 			},
@@ -781,16 +703,12 @@ func Test_ValidateDatacenterFieldChanges(t *testing.T) {
 					Size: 16,
 					Racks: []Rack{{
 						Name: "rack0",
-						Zone: "zone0",
 					}, {
 						Name: "rack1",
-						Zone: "zone1",
 					}, {
 						Name: "rack2",
-						Zone: "zone2",
 					}, {
 						Name: "rack3",
-						Zone: "zone3",
 					}},
 				},
 			},